// packages/transform/src/16/types.ts
export type Node = {
  ParseResult: ParseResult;
// ...
};
export interface ParseResult {
  version?: number;
  stmts?: RawStmt[];
}
// ...
export interface RawStmt {
  stmt?: Node;
  stmt_location?: number;
  stmt_len?: number;
}